chore: revert last commit and remove deprecated polyfill of site
