+++ components/breadcrumb/breadcrumb.module.ts
@@ -13,8 +13,9 @@ import { NzOverlayModule } from 'ng-zorro-antd/core/overlay';
 import { NzDropDownModule } from 'ng-zorro-antd/dropdown';
 import { NzIconModule } from 'ng-zorro-antd/icon';
 
+import { NzBreadCrumbItemComponent } from './breadcrumb-item.component';
 import { NzBreadCrumbSeparatorComponent } from './breadcrumb-separator.component';
-import { NzBreadCrumbComponent, NzBreadCrumbItemComponent } from './breadcrumb.component';
+import { NzBreadCrumbComponent } from './breadcrumb.component';
 
 @NgModule({
   imports: [CommonModule, NzOutletModule, OverlayModule, NzOverlayModule, NzDropDownModule, NzIconModule],
