+++ components/breadcrumb/public-api.ts
@@ -3,6 +3,7 @@
  * found in the LICENSE file at https://github.com/NG-ZORRO/ng-zorro-antd/blob/master/LICENSE
  */
 
+export * from './breadcrumb-item.component';
 export * from './breadcrumb.component';
 export * from './breadcrumb.module';
 export * from './breadcrumb-separator.component';
