+++ package.json
@@ -27,8 +27,7 @@
     "@angular/cdk": "^13.0.1",
     "@ant-design/icons-angular": "^13.0.0-beta.0",
     "date-fns": "^2.16.1",
-    "ngx-hover-preload": "0.0.3",
-    "resize-observer-polyfill": "^1.5.1"
+    "ngx-hover-preload": "0.0.3"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^13.0.2",
